% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice_fun.R
\name{Fn1}
\alias{Fn1}
\title{Fn1 (\strong{internal function})}
\usage{
Fn1(x, foo)
}
\arguments{
\item{x}{\link{numeric} (\strong{required}) variable's values}

\item{foo}{\link{function} (\strong{required}) the function to be applied to the variable x}
}
\value{
the cumulative distribution value of the function foo for the variable x
}
\description{
Cumulative distribution function for a function foo
}
