% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice_fun.R
\name{Fn2}
\alias{Fn2}
\title{Fn2 (\strong{internal function})}
\usage{
Fn2(p, x, foo)
}
\arguments{
\item{p}{\link{numeric} (\strong{required}) probability}

\item{x}{\link{numeric} (\strong{required}) a variable}

\item{foo}{\link{function} (\strong{required}) the function to be applied to the variable x}
}
\value{
a probability
}
\description{
probability for x of density foo
}
