% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice_fun.R
\name{Fn3}
\alias{Fn3}
\title{Fn3 (\strong{internal function})}
\usage{
Fn3(q, x, foo)
}
\arguments{
\item{q}{\link{numeric} (\strong{required}) quantile}

\item{x}{\link{numeric} (\strong{required}) variable's values}

\item{foo}{\link{function} (\strong{required}) the function to be applied to the variable x}
}
\value{
a quantile
}
\description{
quantile q for the foo distribution
}
