% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shrink.R
\name{Shrink}
\alias{Shrink}
\title{Shrink}
\usage{
Shrink(foo, x0, y0, L, R, m = 100, Rmx = 475, Lmin = 20)
}
\arguments{
\item{foo}{\link{function} (\strong{required}) function proportional to the density}

\item{x0}{\link{numeric} (\strong{required}) the current point}

\item{y0}{\link{numeric} (\strong{required}) the vertical level defining the slice}

\item{L}{\link{numeric} (\strong{required}) left boundary of the interval to sample from}

\item{R}{\link{numeric} (\strong{required}) right boundary of the interval to sample from}

\item{m}{\link{numeric} (\strong{with default})}

\item{Rmx}{\link{numeric} (\strong{with default})}

\item{Lmin}{\link{numeric} (\strong{with default})}
}
\value{
a list with the following elements

**Elements Type  Description

$x1\link{numeric} the new point

$L \link{numeric} the new left boundary of the slice

$R \link{numeric} the new right boundary of the slice
}
\description{
Slice sampler: shrinkage procedure
Neal(2003) fig.5
}
