% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice_Init.R
\name{Slice_Init}
\alias{Slice_Init}
\title{Slice_Init}
\usage{
Slice_Init(x, y)
}
\arguments{
\item{x}{\link{numeric} (\strong{required}) a variable}

\item{y}{\link{numeric} (\strong{required}) the measured value for x}
}
\value{
a list with the following elements

$hist_y a histogram of y density

$foo_y a function p(y)

$foo_x a function y(x)
}
\description{
function displaying the measured data (x,y),
and establishing the corresponding functions y(x) and p(y)
}
