% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice_Run.R
\name{Slice_Run}
\alias{Slice_Run}
\title{Slice_Run}
\usage{
Slice_Run(x0, foo_x, foo_y, hist_y, x, w = 20, m = 20)
}
\arguments{
\item{x0}{\link{numeric} (\strong{required}) the current point}

\item{foo_x}{a function y(x)}

\item{foo_y}{a function p(y)}

\item{hist_y}{a histogram of y density}

\item{x}{need?}

\item{w}{\link{numeric} (\strong{required}) estimate of the typical size of a slice}

\item{m}{\link{numeric} (\strong{required}) integer limiting the size of a slice to mw}
}
\value{
a list with the following elements

**Elements Type  Description

$x1\link{numeric} the new point

$L \link{numeric} the new left boundary of the slice

$R \link{numeric} the new right boundary of the slice

$y0\link{numeric} the vertical level defining the slice
}
\description{
function determining the slice and the new point x1
}
