% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice_Cycle.R
\name{Slice_Cycle}
\alias{Slice_Cycle}
\title{Slice_Cycle}
\usage{
Slice_Cycle(x, y, n_iter = 1000, n_burnin = n_iter/2)
}
\arguments{
\item{x}{\link{numeric} (\strong{required}) a variable}

\item{y}{\link{numeric} (\strong{required}) the measured value for x}

\item{n_iter}{\link{numeric} (\strong{with default value}) number of total iterations (including burn in; default: 1000)}

\item{n_burnin}{\link{numeric} (\strong{with default}) length of burn in, i.e. number of iterations to discard at the beginning.
Default is n_iter/2, that is, discarding the first half of the simulations.}
}
\value{
a list with the following elements

**Elements Type  Description

$x1\link{numeric} the new point

$L \link{numeric} left boundary of the slice

$R \link{numeric} right boundary of the slice

$iter \link{numeric}the number of iteration
}
\description{
Slice_Cycle
}
