% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice_mc.R
\name{Slice_mc}
\alias{Slice_mc}
\title{Slice_mc}
\usage{
Slice_mc(x, y, n_iter = 1000, n_burnin = n_iter/2, threshold = min(y[y > 0]))
}
\arguments{
\item{x}{\link{numeric} (\strong{required}) a variable}

\item{y}{\link{numeric} (\strong{required}) the measured value for x}

\item{n_iter}{\link{numeric} (\strong{with default value}) number of total iterations (including burn in; default: 1000)}

\item{n_burnin}{\link{numeric} (\strong{with default}) length of burn in, i.e. number of iterations to discard at the beginning.
Default is n_iter/2, that is, discarding the first half of the simulations.}

\item{threshold}{\link{numeric} (\strong{with default}) measured point below or equal to the threshold are rejected.
Default is the minimum measured value.}
}
\value{
a list with the following elements
\tabular{lll}{
\strong{Element} \tab \strong{Type} \tab \strong{Description}\cr
\verb{$x1} \tab \code{numeric} \tab the new point \cr
\verb{$L} \tab \code{numeric} \tab the left boundary of the slice \cr
\verb{$R} \tab \code{numeric} \tab the right boundary of the slice \cr
}
}
\description{
Monte-Carlo procedure
}
\examples{
#TL example from RLumModel
require(RLumModel)
data("ExampleData.ModelOutput", envir = environment())
TL_curve <- get_RLum(model.output, recordType = "TL$", drop = FALSE)
x<-TL_curve$data[,1]
y<-TL_curve$data[,2]
if (dev.cur()!=1) dev.off()
Slice_mc(x,y,n_iter=1000,n_burnin=500)


}
