% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice_Run.R
\name{Slice_Run}
\alias{Slice_Run}
\title{Slice_Run}
\usage{
Slice_Run(x0, foo_x, foo_y, hist_y, w = 20, m = 20)
}
\arguments{
\item{x0}{\link{numeric} (\strong{required}) the current point}

\item{foo_x}{a function y(x)}

\item{foo_y}{a function p(y)}

\item{hist_y}{a histogram of y density}

\item{w}{\link{numeric} (\strong{required}) estimate of the typical size of a slice}

\item{m}{\link{numeric} (\strong{required}) integer limiting the size of a slice to mw}
}
\value{
a list with the following elements
\tabular{lll}{
\strong{Element} \tab \strong{Type} \tab \strong{Description}\cr
\verb{$x1} \tab \code{numeric} \tab the new point \cr
\verb{$L} \tab \code{numeric} \tab the new left boundary of the slice \cr
\verb{$R} \tab \code{numeric} \tab the new right boundary of the slice \cr
\verb{$y0} \tab \code{numeric} \tab the vertical level defining the slice\cr
}
}
\description{
function determining the slice and the new point x1
}
\examples{
#TL example from RLumModel
require(RLumModel)
data("ExampleData.ModelOutput", envir = environment())
TL_curve <- get_RLum(model.output, recordType = "TL$", drop = FALSE)
x<-TL_curve$data[,1]
y<-TL_curve$data[,2]
x0<-300
dev.off()
run0<-Slice_Init(x,y)
Slice_Run(x0,run0$foo_x,run0$foo_y,run0$hist_y)

#multiples TL calculées avec RLumModel##########
#call function "model_LuminescenceSignals", model = "Bailey2001"
# the irradiation dose is varied and then compared.
irradiation_dose <- seq(from = 0,to = 100,by = 20)
model.output <- lapply(irradiation_dose,
      function(x){
           sequence <- list(IRR = c(20, x, 1),
              #PH = c(220, 10, 5),
              TL=c(20,400,5))
           data <- model_LuminescenceSignals(
              sequence = sequence,
              model = "Bailey2001",
              plot = FALSE,
              verbose = FALSE)
           return(get_RLum(data, recordType = "TL$", drop = FALSE))
})
##combine output curves
TL_curve.merged <- merge_RLum(model.output)
##plot
plot_RLum(
 object = TL_curve.merged,
 xlab = "Temperature [°C]",
 ylab = "TL signal [a.u.]",
 main = "TL signal with various dose",
 legend.text = paste("dose", irradiation_dose, "Gy"),
 combine = TRUE)
##
n.pt<-length(TL_curve.merged[1]$data[,1])
n.irr<-length(irradiation_dose)
y<-x<-array(dim=c(n.pt,n.irr))
for (i in 1:n.irr){
 x[,i]<-TL_curve.merged[i]$data[,1]
 y[,i]<-TL_curve.merged[i]$data[,2]
}

x0<-300
run0<-Slice_Init(x[,1],y[,1])
Slice_Run(x0,run0$foo_x,run0$foo_y,run0$hist_y)

}
